requirements.txt
# LangChain Core
langchain
langchain-core

# # OpenAI Integration
# langchain-openai
# openai

# # Anthropic Integration
# langchain-anthropic

# # Google Gemini (PaLM) Integration
# langchain-google-genai
# google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
# numpy
# scikit-learn

## pytorch for Hugging Face models
# torch
# torchvision

## For sentence embeddings
# sentence-transformers

## Streamlit for building interactive applications
streamlit

## For using Langchan-GROQ
langchain-groq
langchain_huggingface

# Model ID	Size	RAM Needed
# TinyLlama/TinyLlama-1.1B-Chat-v1.0	1.1B	~2 GB
# Qwen/Qwen2.5-1.5B-Instruct	1.5B	~3 GB
# microsoft/phi-2	2.7B	~6 GB
# google/gemma-3-1b-it	1B	~2 GB
# HuggingFaceTB/SmolLM2-1.7B-Instruct	1.7B	~4 GB